move value cache to star tree context + other comments
