+++ server/src/main/java/org/opensearch/search/internal/SearchContext.java
@@ -31,12 +31,10 @@
 
 package org.opensearch.search.internal;
 
-import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.search.Collector;
 import org.apache.lucene.search.CollectorManager;
 import org.apache.lucene.search.FieldDoc;
 import org.apache.lucene.search.Query;
-import org.apache.lucene.util.FixedBitSet;
 import org.opensearch.action.search.SearchShardTask;
 import org.opensearch.action.search.SearchType;
 import org.opensearch.common.Nullable;
@@ -127,7 +125,6 @@ public abstract class SearchContext implements Releasable {
     private final List<Releasable> releasables = new CopyOnWriteArrayList<>();
     private final AtomicBoolean closed = new AtomicBoolean(false);
     private InnerHitsContext innerHitsContext;
-    protected volatile Map<LeafReaderContext, FixedBitSet> starTreeValuesMap;
     private volatile boolean searchTimedOut;
     private StarTreeQueryContext starTreeQueryContext;
 
@@ -542,13 +539,5 @@ public abstract class SearchContext implements Releasable {
 
     public StarTreeQueryContext getStarTreeQueryContext() {
         return this.starTreeQueryContext;
-    }
-
-    public void initializeStarTreeValuesMap() {
-        this.starTreeValuesMap = new HashMap<>();
-    }
-
-    public Map<LeafReaderContext, FixedBitSet> getStarTreeValuesMap() {
-        return starTreeValuesMap;
     }
 }
