+++ server/src/main/java/org/opensearch/search/startree/StarTreeQueryContext.java
@@ -8,10 +8,13 @@
 
 package org.opensearch.search.startree;
 
+import org.apache.lucene.index.LeafReaderContext;
+import org.apache.lucene.util.FixedBitSet;
 import org.opensearch.common.annotation.ExperimentalApi;
 import org.opensearch.index.codec.composite.CompositeIndexFieldInfo;
 
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Query class for querying star tree data structure.
@@ -31,23 +34,21 @@ public class StarTreeQueryContext {
      * Map of field name to a value to be queried for that field
      * This is used to filter the data based on the query
      */
-    private final Map<String, Long> queryMap;
+    private volatile Map<String, Long> queryMap;
 
-    // /**
-    // * Cache for leaf results
-    // * This is used to cache the results for each leaf reader context
-    // * to avoid reading the data from the leaf reader context multiple times
-    // */
-    // private volatile Map<LeafReaderContext, Map<String, StarTreeQueryHelper.MetricInfo>> leafResultsCache;
+    /**
+    * Cache for leaf results
+    * This is used to cache the results for each leaf reader context
+    * to avoid reading the filtered values from the leaf reader context multiple times
+    */
+    protected volatile Map<LeafReaderContext, FixedBitSet> starTreeValuesMap;
 
-    // /**
-    // * List of metrics to be computed & cached
-    // */
-    // private List<StarTreeQueryHelper.MetricInfo> metrics;
-
-    public StarTreeQueryContext(CompositeIndexFieldInfo starTree, Map<String, Long> queryMap) {
+    public StarTreeQueryContext(CompositeIndexFieldInfo starTree, Map<String, Long> queryMap, boolean cacheStarTreeValues) {
         this.starTree = starTree;
         this.queryMap = queryMap;
+        if (cacheStarTreeValues) {
+            starTreeValuesMap = new ConcurrentHashMap<>();
+        }
     }
 
     public CompositeIndexFieldInfo getStarTree() {
@@ -56,5 +57,9 @@ public class StarTreeQueryContext {
 
     public Map<String, Long> getQueryMap() {
         return queryMap;
+    }
+
+    public Map<LeafReaderContext, FixedBitSet> getStarTreeValuesMap() {
+        return starTreeValuesMap;
     }
 }
